# src/main/resources/openapi/openapi.yaml
openapi: 3.0.3

info:
  title: Forum API
  description: REST API for forum topics (posts), replies, and users (JWT bearer auth).
  version: 1.0.0

servers:
  - url: http://localhost:9000
    description: Local

tags:
  - name: Auth
    description: Authentication and registration
  - name: Posts
    description: Topic (post) operations
  - name: Replies
    description: Replies on topics
  - name: Users
    description: User operations

paths:

  /auth/login:
    post:
      tags: [Auth]
      operationId: login
      summary: Obtain a JWT (username + password)
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
        "401":
          description: Invalid credentials

  /auth/register:
    post:
      tags: [Auth]
      operationId: register
      summary: Register a new user (role USER)
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserResponse"
        "409":
          description: Username or email already in use

  /posts:
    get:
      tags: [Posts]
      operationId: listPosts
      summary: List all topics
      security: []
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PostResponse"

    post:
      tags: [Posts]
      operationId: createPost
      summary: Create a topic
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePostRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PostResponse"
        "409":
          description: Title already in use

  /posts/{id}:
    get:
      tags: [Posts]
      operationId: getPostById
      summary: Get topic by id (increments view count by default)
      security: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
        - name: trackView
          in: query
          required: false
          description: When false, do not increment view count.
          schema:
            type: boolean
            default: true
      responses:
        "200":
          description: Found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PostResponse"
        "404":
          description: Not found

    put:
      tags: [Posts]
      operationId: updatePost
      summary: Update a topic
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdatePostRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PostResponse"
        "403":
          description: Forbidden
        "404":
          description: Not found
        "409":
          description: Title already in use

    delete:
      tags: [Posts]
      operationId: deletePost
      summary: Delete a topic
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "204":
          description: No content
        "403":
          description: Forbidden
        "404":
          description: Not found

  /posts/{postId}/replies:
    get:
      tags: [Replies]
      operationId: listRepliesForPost
      summary: List replies for a topic (paginated, 10 per page by default)
      security: []
      parameters:
        - name: postId
          in: path
          required: true
          schema:
            type: integer
            format: int64
        - name: page
          in: query
          required: false
          schema:
            type: integer
            minimum: 0
            default: 0
        - name: size
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 10
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReplyPageResponse"
        "404":
          description: Topic not found

    post:
      tags: [Replies]
      operationId: createReply
      summary: Create a reply on a topic
      parameters:
        - name: postId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateReplyRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReplyResponse"
        "404":
          description: Topic not found

  /replies/{id}:
    get:
      tags: [Replies]
      operationId: getReplyById
      summary: Get reply by id
      security: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: Found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReplyResponse"
        "404":
          description: Not found

    put:
      tags: [Replies]
      operationId: updateReply
      summary: Update a reply
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateReplyRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReplyResponse"
        "403":
          description: Forbidden
        "404":
          description: Not found

    delete:
      tags: [Replies]
      operationId: deleteReply
      summary: Delete a reply
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "204":
          description: No content
        "403":
          description: Forbidden
        "404":
          description: Not found

  /users:
    get:
      tags: [Users]
      operationId: listUsers
      summary: List all users
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/UserResponse"

    post:
      tags: [Users]
      operationId: createUser
      summary: Create a user (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateUserRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserResponse"
        "409":
          description: Username or email already in use

  /users/{id}:
    get:
      tags: [Users]
      operationId: getUserById
      summary: Get user by id
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: Found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserResponse"
        "404":
          description: Not found

    put:
      tags: [Users]
      operationId: updateUser
      summary: Update a user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserResponse"
        "403":
          description: Forbidden (e.g. only admin can assign MODERATOR)
        "404":
          description: Not found
        "409":
          description: Username or email already in use

    delete:
      tags: [Users]
      operationId: deleteUser
      summary: Delete a user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "204":
          description: No content
        "404":
          description: Not found

components:

  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    LoginRequest:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
          minLength: 1
          maxLength: 100
        password:
          type: string
          minLength: 1
          maxLength: 128

    LoginResponse:
      type: object
      required: [accessToken, tokenType, expiresInSeconds]
      properties:
        accessToken:
          type: string
        tokenType:
          type: string
          example: Bearer
        expiresInSeconds:
          type: integer
          format: int64

    RegisterRequest:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
          minLength: 1
          maxLength: 100
        email:
          type: string
          format: email
          maxLength: 320
        password:
          type: string
          minLength: 8
          maxLength: 72
          description: Stored as a bcrypt hash; max 72 bytes for bcrypt.

    UserRole:
      type: string
      description: USER denotes a regular (non-privileged) account.
      enum: [ADMIN, MODERATOR, USER]

    CreateUserRequest:
      type: object
      required: [username, role, password]
      properties:
        username:
          type: string
          minLength: 1
          maxLength: 100
        email:
          type: string
          format: email
          maxLength: 320
        role:
          $ref: "#/components/schemas/UserRole"
        password:
          type: string
          minLength: 8
          maxLength: 72
          description: Stored as a bcrypt hash; max 72 bytes for bcrypt.

    UpdateUserRequest:
      type: object
      required: [username, role]
      properties:
        username:
          type: string
          minLength: 1
          maxLength: 100
        email:
          type: string
          format: email
          maxLength: 320
          nullable: true
        role:
          $ref: "#/components/schemas/UserRole"
        password:
          type: string
          minLength: 8
          maxLength: 72
          description: If set, replaces the password.

    UserResponse:
      type: object
      required: [id, username, role, createdAt]
      properties:
        id:
          type: integer
          format: int64
        username:
          type: string
        email:
          type: string
          format: email
          nullable: true
        role:
          $ref: "#/components/schemas/UserRole"
        createdAt:
          type: string
          format: date-time

    AuthorSummary:
      type: object
      required: [id, username]
      properties:
        id:
          type: integer
          format: int64
        username:
          type: string

    CreatePostRequest:
      type: object
      required: [title, content]
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 500
        content:
          type: string
          maxLength: 10000

    UpdatePostRequest:
      type: object
      required: [title, content]
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 500
        content:
          type: string
          maxLength: 10000

    PostResponse:
      type: object
      required: [id, title, content, author, createdAt, updatedAt, viewCount]
      properties:
        id:
          type: integer
          format: int64
        title:
          type: string
        content:
          type: string
        author:
          $ref: "#/components/schemas/AuthorSummary"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        viewCount:
          type: integer
          format: int64

    CreateReplyRequest:
      type: object
      required: [content]
      properties:
        content:
          type: string
          maxLength: 10000

    UpdateReplyRequest:
      type: object
      required: [content]
      properties:
        content:
          type: string
          maxLength: 10000

    ReplyResponse:
      type: object
      required: [id, postId, content, author, createdAt, updatedAt]
      properties:
        id:
          type: integer
          format: int64
        postId:
          type: integer
          format: int64
        content:
          type: string
        author:
          $ref: "#/components/schemas/AuthorSummary"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time

    ReplyPageResponse:
      type: object
      required: [content, page, size, totalElements, totalPages]
      properties:
        content:
          type: array
          items:
            $ref: "#/components/schemas/ReplyResponse"
        page:
          type: integer
        size:
          type: integer
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer

security:
  - bearerAuth: []